refactor: common/lib -> common/api directory change
